requirements.txt
aiotdlib==0.27.6
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
fastapi==0.120.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.2.1
python-telegram-bot==22.5
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.38.0
